HTTP Communication: Tentative getAllQuestions

+++ client-mobile/app/src/main/java/com/example/polyquiz/Question.kt
@@ -4,7 +4,7 @@ data class Question(
     val id: String,
     val type: String,
     val text: String,
-    val points: Int,
+    val points: Double,
     val choices: List<Choice>? = null,
     val answer: String? = null,
     val lastModification: String
